removing the ability to add duplicate reviews, adding an update review and delete review functionalities

+++ BackEnd - Spring/MovieLovers/src/main/kotlin/com/MichaelRichards/MovieLovers/dtos/MovieReviewDTO.kt
@@ -2,6 +2,6 @@ package com.MichaelRichards.MovieLovers.dtos
 
 data class MovieReviewDTO(
     val imdbId: String,
-    val rating: Int,
+    val rating: Float,
     val description: String,
 )
